refactor(Dropdown): Use `partials` pattern
