+++ packages/react-component-library/src/components/Dropdown/partials/StyledAdornment.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const StyledAdornment = styled.span`
+  display: flex;
+  align-items: center;
+`
