+++ packages/react-component-library/src/components/Dropdown/partials/StyledEndAdornment.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { spacing } = selectors
+
+export const StyledEndAdornment = styled.span`
+  position: absolute;
+  top: ${spacing('px')};
+  right: 0;
+`
