+++ packages/react-component-library/src/components/Dropdown/partials/StyledIcon.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components'
+import { selectors } from '@royalnavy/design-tokens'
+
+const { spacing } = selectors
+
+export const StyledIcon = styled.span`
+  position: relative;
+  top: ${spacing('px')};
+`
